Initialization of anonymous bit-fields in structs
For anonymous bit-fields in structs, carrier fields may
be introduced which are not initialized since no default
initializer are generated earlier. This cause the translation in
C2C to throw an error since too few initializers are available.

Example:

struct s2 {
		int : 10;
		int a;
		int : 10;
		char b;
		int : 10;
		short c;
		int : 10;
} s2 = { 42, 'a', 43 };

To work around this issue we need to generate a default inializer
for every new member that does not have a translated member.

Based on P#80, with more efficient algorithms.

Bug 23362