When testing builtin functions, prevent constant propagation
Now that some builtin functions have known semantics, constant
propagation can happen in this test.  This defeats the purpose,
which is to check that the correct processor instructions are generated.

To prevent this constant propagation, we move the initialized
variables to global scope.  Since they are not "const", their
values are not known to the optimizer.