clightgen: handle empty names given to padding bit fields
In the Clight AST, padding bit fields (such as `int : 6;`) in composite
declarations are given an ident that corresponds to the empty string.

Previously, clightgen would give name `_` to this ident, but this is
not valid Coq.

This commit gives name `empty_ident` to the empty ident.  This name
does not start with an underscore, so it cannot conflict with the
names for regular idents, which all start with `_`.